Moved Tests Into Sub-Folders Sorted test according to their respective place in Lib & updated docs
