+++ __Tests__/Utils/Indent.test.ts
@@ -1,4 +1,4 @@
-import Spc from '../Lib/Utils/Indent';
+import Spc from '../../Lib/Utils/Indent';
 
 describe('Indent', () => {
 	it('Should leave space equivalent to the number then attach the string given', () => {
